WIP... dial in midi channels a lil
