+++ src/components/ConfigModal.tsx
@@ -1,9 +1,23 @@
-import { Modal, Stack, Group, Title, Divider, Flex, Button } from "@mantine/core";
+import {
+  Modal,
+  Stack,
+  Group,
+  Title,
+  Divider,
+  Flex,
+  Button,
+  Fieldset,
+  Select,
+} from "@mantine/core";
 import { IconRefresh, IconWorldDown, IconWorldUp } from "@tabler/icons-react";
 import { ConfigPanel } from "./ConfigPanel";
 import { JunoButtonGroup } from "./JunoButtonGroup";
 import { useKiwi106Context } from "../hooks/useKiwi106Context";
 import { useMidiContext } from "../hooks/useMidiContext";
+import { isMidiChannel } from "../types/Midi";
+import { SelectMidiChannel } from "./SelectMidiChannel";
+import { MidiPortData } from "../contexts/MidiContext";
+import { useConfigStore } from "../stores/configStore";
 
 interface ConfigModalProps {
   opened: boolean;
@@ -12,19 +26,23 @@ interface ConfigModalProps {
 
 export const ConfigModal = ({ opened, onClose }: ConfigModalProps) => {
   const midiContext = useMidiContext();
+  const configStore = useConfigStore();
   const { kiwiMidi } = useKiwi106Context();
-  
+
+  const formatName = (i: MidiPortData | null) =>
+    i == null ? null : `${i.manufacturer} ${i.name}`;
+  const findInputByFormattedName = (fn: string | null) =>
+    configStore.availableInputs.find((i) => formatName(i) === fn) ?? null;
+  const findOutputByFormattedName = (fn: string | null) =>
+    configStore.availableOutputs.find((o) => formatName(o) === fn) ?? null;
+
   const sendSysexGlobalDump = () => {
     // TODO: grab all the shit
-    console.log("[sendSysexGlobalDump]")
-  }
+    console.log("[sendSysexGlobalDump]");
+  };
 
   return (
-    <Modal.Root
-      opened={opened}
-      onClose={onClose}
-      size="xl"
-    >
+    <Modal.Root opened={opened} onClose={onClose} size="xl">
       <Modal.Overlay />
       <Modal.Content>
         <Modal.Header px={0} pb={4}>
@@ -66,7 +84,69 @@ export const ConfigModal = ({ opened, onClose }: ConfigModalProps) => {
           </Stack>
         </Modal.Header>
         <Modal.Body pt="md">
-          <ConfigPanel />
+          <Stack>
+            <Group>
+              <Fieldset legend="Input">
+                <Group wrap="nowrap">
+                  <Select
+                    label="Device"
+                    clearable
+                    allowDeselect={false}
+                    placeholder={
+                      midiContext.enabled ? "Select an input" : "Not available"
+                    }
+                    value={formatName(configStore.input)}
+                    data={configStore.availableInputs.map(
+                      (i) => `${i.manufacturer} ${i.name}`
+                    )}
+                    onChange={(fn) =>
+                      configStore.setInput(findInputByFormattedName(fn))
+                    }
+                  ></Select>
+                  <SelectMidiChannel
+                    enabled={midiContext.enabled}
+                    value={configStore.inputChannel}
+                    onChange={(c) => {
+                      if (isMidiChannel(c)) {
+                        configStore.setInputChannel(c);
+                      }
+                    }}
+                  />
+                </Group>
+              </Fieldset>
+
+              <Fieldset legend="Output">
+                <Group wrap="nowrap">
+                  <Select
+                    label="Device"
+                    clearable
+                    allowDeselect={false}
+                    placeholder={
+                      midiContext.enabled ? "Select an output" : "Not available"
+                    }
+                    value={formatName(configStore.output)}
+                    data={configStore.availableOutputs.map(
+                      (i) => `${i.manufacturer} ${i.name}`
+                    )}
+                    onChange={(fn) =>
+                      configStore.setOutput(findOutputByFormattedName(fn))
+                    }
+                  ></Select>
+                  <SelectMidiChannel
+                    enabled={midiContext.enabled}
+                    value={configStore.outputChannel}
+                    onChange={(c) => {
+                      if (isMidiChannel(c)) {
+                        configStore.setOutputChannel(c);
+                      }
+                    }}
+                  />
+                </Group>
+              </Fieldset>
+            </Group>
+
+            <ConfigPanel />
+          </Stack>
         </Modal.Body>
       </Modal.Content>
     </Modal.Root>
