+++ src/components/ConfigPanel.tsx
@@ -1,6 +1,4 @@
 import {
-  Button,
-  Fieldset,
   Group,
   List,
   ListItem,
@@ -8,93 +6,18 @@ import {
   Stack,
   Text,
 } from "@mantine/core";
-import { useMidiContext } from "../hooks/useMidiContext";
-import { MidiPortData } from "../contexts/MidiContext";
-import { SelectMidiChannel } from "./SelectMidiChannel";
-import { IconRefresh, IconWorldDown, IconWorldUp } from "@tabler/icons-react";
-import { useConfigStore } from "../stores/configStore";
-import { isMidiChannel } from "../types/Midi";
 import { useKiwi106Context } from "../hooks/useKiwi106Context";
 import { Selector } from "./Selector";
 import { kiwi106MessageModes } from "../types/KiwiGlobalData";
+import { MidiChannel } from "../types/Midi";
 
 export const ConfigPanel = () => {
-  const midiContext = useMidiContext();
   const kiwi106Context = useKiwi106Context();
-  const configStore = useConfigStore();
-
-  const formatName = (i: MidiPortData | null) =>
-    i == null ? null : `${i.manufacturer} ${i.name}`;
-  const findInputByFormattedName = (fn: string | null) =>
-    configStore.availableInputs.find((i) => formatName(i) === fn) ?? null;
-  const findOutputByFormattedName = (fn: string | null) =>
-    configStore.availableOutputs.find((o) => formatName(o) === fn) ?? null;
-
-  const requestGlobalDumpSysex = () => {
-    kiwi106Context.kiwiMidi?.requestSysexGlobalDump();
-  };
-
+  
+  const midiChannel: MidiChannel = 16;
+  console.log("Midichannel", midiChannel)
   return (
     <Stack>
-      <Group>
-        <Fieldset legend="Input">
-          <Group wrap="nowrap">
-            <Select
-              label="Device"
-              clearable
-              allowDeselect={false}
-              placeholder={
-                midiContext.enabled ? "Select an input" : "Not available"
-              }
-              value={formatName(configStore.input)}
-              data={configStore.availableInputs.map(
-                (i) => `${i.manufacturer} ${i.name}`
-              )}
-              onChange={(fn) =>
-                configStore.setInput(findInputByFormattedName(fn))
-              }
-            ></Select>
-            <SelectMidiChannel
-              enabled={midiContext.enabled}
-              value={configStore.inputChannel}
-              onChange={(c) => {
-                if (isMidiChannel(c)) {
-                  configStore.setInputChannel(c);
-                }
-              }}
-            />
-          </Group>
-        </Fieldset>
-
-        <Fieldset legend="Output">
-          <Group wrap="nowrap">
-            <Select
-              label="Device"
-              clearable
-              allowDeselect={false}
-              placeholder={
-                midiContext.enabled ? "Select an output" : "Not available"
-              }
-              value={formatName(configStore.output)}
-              data={configStore.availableOutputs.map(
-                (i) => `${i.manufacturer} ${i.name}`
-              )}
-              onChange={(fn) =>
-                configStore.setOutput(findOutputByFormattedName(fn))
-              }
-            ></Select>
-            <SelectMidiChannel
-              enabled={midiContext.enabled}
-              value={configStore.outputChannel}
-              onChange={(c) => {
-                if (isMidiChannel(c)) {
-                  configStore.setOutputChannel(c);
-                }
-              }}
-            />
-          </Group>
-        </Fieldset>
-      </Group>
       <Stack>
         <List>
           <ListItem>Program version {kiwi106Context.programVersion}</ListItem>
