+++ src/types/IntRange.ts
@@ -5,6 +5,8 @@ type Enumerate<
   ? Acc[number]
   : Enumerate<N, [...Acc, Acc["length"]]>;
 
+// List of values [n,m)
+// so [0, 4] would equate to 0 | 1 | 2 | 3
 export type IntRange<F extends number, T extends number> = Exclude<
   Enumerate<T>,
   Enumerate<F>
