+++ src/utils/trimMidiCcValue.ts
@@ -1,7 +1,12 @@
-import { MidiCcValue, MidiChannel } from "../types/Midi";
+import { isMidiChannel, MidiCcValue, MidiChannel } from "../types/Midi";
 
-export const trimMidiChannel: (n: number) => MidiChannel = (v) => {
-  return Math.min(15, Math.max(0, v)) as MidiChannel;
+export const trimMidiChannel: (v: number) => MidiChannel = (v) => {
+  const midiChannel = Math.min(15, Math.max(0, v)) + 1;
+  if (isMidiChannel(midiChannel)) {
+    return midiChannel;
+  } else {
+    throw new Error(`Invalid MidiChannel value (${v})`);
+  }
 };
 
 export const trimMidiCcValue: (n: number) => MidiCcValue = (v) => {
